consolidate Kiota enums into the module
