+++ vscode/microsoft-kiota/src/kiotaInterop/connect.ts
@@ -1,5 +1,6 @@
 import * as cp from 'child_process';
 import * as rpc from 'vscode-jsonrpc/node';
+
 import { getKiotaPath, ensureKiotaIsPresent } from './install';
 
 export default async function connectToKiota<T>(callback: (connection: rpc.MessageConnection) => Promise<T | undefined>, workingDirectory: string = process.cwd()): Promise<T | undefined> {
