+++ vscode/microsoft-kiota/src/kiotaInterop/enums.ts
@@ -1,0 +1,31 @@
+export enum KiotaGenerationLanguage {
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  CSharp = 0,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  Java = 1,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  TypeScript = 2,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  PHP = 3,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  Python = 4,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  Go = 5,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  Swift = 6,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  Ruby = 7,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  CLI = 8,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  Dart = 9,
+};
+
+export enum KiotaPluginType {
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  OpenAI = 0,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  ApiManifest = 1,
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  ApiPlugin = 2,
+};
